serverless quits before uploading assets
I believe the promise isn't being returned properly for the `putObject` request. Nothing is returned from the `glob.sync().forEach` function. This seems to be causing the serverless CLI to exist too quickly before the files have even uploaded. I changed it to a BbPromise.each and now it is working for me.